requirements.txt
fastapi==0.115.8
neo4j==5.27.0
passlib==1.7.4
pydantic==2.10.6
pydantic_extra_types==2.10.2
pydantic_settings==2.7.1
PyJWT==2.10.1
uvicorn==0.34.0
twilio==9.4.4
phonenumbers==8.13.54
python-multipart==0.0.12
uvicorn==0.34.0